remove dev page
