+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    // "strictNullChecks": true,
     "allowJs": false,
     "allowSyntheticDefaultImports": true,
     "baseUrl": "./src",
@@ -15,7 +16,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "target": "ES2017"
-    // "strictNullChecks": true
   },
   "include": ["src", "typings/*.d.ts"],
   "files": ["typings/global.d.ts"],
